+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 - Fix dot completion issue with React primitives. https://github.com/rescript-lang/rescript/pull/7292
 - Stdlib namespace for Core modules (fixes name clashes with user modules). https://github.com/rescript-lang/rescript/pull/7285
 - Fix runtime type check for Object in untagged variants when one variant case is `null`. https://github.com/rescript-lang/rescript/pull/7303
+- Fix files that were being truncated when sent to the CDN over FTP. https://github.com/rescript-lang/rescript/pull/7306
 
 #### :house: Internal